move export_to_pandas to results_interface
