+++ src/powfacpy/results_interface.py
@@ -2,7 +2,8 @@ import sys
 sys.path.insert(0,r'.\src')
 import powfacpy
 import numpy as np
-
+from os import remove, getcwd
+from pandas import read_csv
 
 class PFResultsInterface(powfacpy.PFBaseInterface):
 
@@ -28,6 +29,28 @@ class PFResultsInterface(powfacpy.PFBaseInterface):
     list = intvec.V
     intvec.Delete()
     return list
+  
+  def export_to_pandas(self, result_objects, elements, variables):
+    """returns pandas DataFrame of specified simulation results.
+      Arguments:
+        result_objects: List of ElmRes for each result variable (repeat if identical or several variables) 
+        elements: List of Powerfactory Objects for each result variable 
+        variables: List of variable names for each result variable
+    """
+    FILE_NAME = 'temp'
+    FILE_PATH = getcwd()
+    FULL_PATH = FILE_PATH + "\\" + FILE_NAME + ".csv"
+    self.export_to_csv(
+      dir= FILE_PATH,
+      file_name=FILE_NAME,
+      results_variables_lists={
+          'result_objects':result_objects,
+          'elements':elements,
+          'variables':variables})
+
+    df = read_csv(FULL_PATH, encoding='ISO-8859-1')
+    remove(FULL_PATH)
+    return df
